% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CablePred.R
\name{CablePred}
\alias{CablePred}
\title{Predict the amount of cable required to reach a desired headrope depth
with a trawl.}
\usage{
CablePred(desired.depth)
}
\arguments{
\item{desired.depth}{A numeric scalar providing the desired head rope depth in meters (also known as
fishing depth).}

\item{rdat}{A character scalar giving the name of the RData file that contains the cable length
and fishing depth data that will be used to fit a model (currently uses lm())
with fishing depth as the predictor and the cable length as the predicted.}
}
\value{
The function generates a plot of fishing depth versus cable lenght and returns estimate
 of the cable length required to achieve a desired fishing depth to the console.
}
\description{
Predict cable required (in feet) to reach a headrope depth (in meters) with a midwater
trawl
}
\details{
This function currently is intended for use with trawl data collected using a sensor
 that measures the depth of the net
 on the USGS RV Sturgeon using a 50 ft midwater trawl. The modeled cable length
 only includes one predictor variable at present. Given more data and more variable
 conditions, additional predictors as well as types of models may be added or
 required.
}
