% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishDepthTemp.R
\name{FishDepthTemp}
\alias{FishDepthTemp}
\title{Create an interactive graph of depth and temperature data recorded by a Marport
 Trawl Explorer or TE/Catch sensor deployed on a trawl. The user can draw a box around a subset
 of data and obtain the mean depth, start time, end time, and elapsed time of that
 subset.}
\usage{
FishDepthTemp()
}
\value{
An interactive graph and metadata corresponding to the data selected.
}
\description{
Create an interactive graph of depth and temperature data recorded by a Marport
 Trawl Explorer or TE/Catch sensor deployed on a trawl. The user can draw a box around a subset
 of data and obtain the mean depth, start time, end time, and elapsed time of that
 subset.
}
\details{
The purpose of this function is to allow the user to graphically select a subset of
 the depth and temperature data recorded by an instrument deployed on the headline of a trawl.
}
