% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxFluorDepth.R
\name{MaxFluorDepth}
\alias{MaxFluorDepth}
\title{Create an interactive graph of CTD data recorded by a Seabird
 The user can draw a box around a subset of data and obtain maximum
 fluorescence depth. This will eventually be modified to allow other variables
 and to save the subsetted data to a csv, allowing one to filter bad data.}
\usage{
MaxFluorDepth()
}
\value{
A pair of graphs with the depth of maximum fluorescence displayed.
}
\description{
Create an interactive graph of CTD data recorded by a Seabird
 The user can draw a box around a subset of data and obtain maximum
 fluorescence depth. This will eventually be modified to allow other variables
 and to save the subsetted data to a csv, allowing one to filter bad data.
}
\details{
The purpose of this function is to allow the user to graphically select a subset of
 CTD data to locate the depth of maximum fluorescence. This could be modified
 to also eliminate erroneous data and save the filtered data.
}
