% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewData.R
\name{NewData}
\alias{NewData}
\title{Input new values for fishing depth and cable length used in midwater trawling
for use in modeling the cable length required to reach a desired fishing depth.
Will prompt user for innformation about the trawl operations.}
\usage{
NewData()
}
\value{
Adds a data pair to the trawl cable length and fishing depth file selected. Creates
a regression plot of cable length versus fishing depth. Creates a plot
of residuals from the regression.
}
\description{
Input new values for fishing depth and cable length used in midwater trawling
for use in modeling the cable length required to reach a desired fishing depth.
}
\details{
#' @param rdat
A character scalar giving the name of the RData file in \code{CablePred}
to which new data will be added and which is used by \code{\link{CablePred}}.


 This function asks the user for new trawl cable length and fishing depth
 values and saves them to the user-specific csv file.
}
