% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSeabird.R
\name{ProcessSeabird}
\alias{ProcessSeabird}
\title{Process Seabird cnv files}
\usage{
ProcessSeabird()
}
\value{
A graph of water temperature and fluorescence profiles as well as
a bin-averaged csv file.
}
\description{
CTD processing from cnv files
}
\details{
Reads a user-selected cnv file and carries out several operations. These
include data editing, bin averaging in 1-m increments, calculation of
surface temperature, thermocline depth, metalimnion depths, and water column
temperature.
}
