% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectTrawlData.R
\name{SelectTrawlData}
\alias{SelectTrawlData}
\title{This function is called by TimeOnBottom(). It provides the interactive data selection
and generation of metadata to be recorded with the selected data.}
\usage{
SelectTrawlData()
}
\value{
A file path and parameters of the data contained in the file at that path.
}
\description{
This function is called by TimeOnBottom(). It provides the interactive data selection
and generation of metadata to be recorded with the selected data.
}
\details{
This function currently works for/is intended for midwater trawl data collected
 on the USGS RV Sturgeon using a 50 ft midwater trawl. The modeled cable length
 only includes one predictor variable at present. Given more data and more variable
 conditions, additional predictors as well as types of models may be added or
 required.
}
