% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors2Data-package.R
\docType{package}
\name{Sensors2Data}
\alias{Sensors2Data}
\alias{Sensors2Data-package}
\title{A Collection of Functions for Use on USGS Fishery Surveys}
\description{
\pkg{Sensors2Data} is a collection of functions for generating data from sensors
or making predictions from those data.
}
\details{
\code{\link{CablePred}} to predict trawl cable required to fish a certain depth,
\code{\link{ProcessSeabird}} to view and save CTD data, and
\code{\link{NewData}} to add measured fishing depth and cable length pairs
for trawling.
}
