% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CablePred.R
\name{CablePred}
\alias{CablePred}
\title{Predict the amount of cable required to reach a desired headrope depth
with a trawl.}
\usage{
CablePred(desired.depth = 10)
}
\arguments{
\item{desired.depth}{Numeric scale the user enters after being prompted which is equal to the desired
fishing depth in meters.}
}
\value{
The function generates a regression plot of fishing depth versus cable length, a plot
of regression residuals versus predicted fishing depth. Also returns an estimate
of the cable length required and the confidence interval to the console.
}
\description{
Predict the amount of cable required to reach a desired headrope depth
with a trawl.
}
\details{
This function requires input data in the form of a csv file with a variable called
Cable.feet and another called Depth.meters. The user will be prompted to lcoate this
file by the function, then asked to input a desired fishing depth (desired.depth)
}
