% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors2Data-package.R
\docType{package}
\name{Sensors2Data}
\alias{Sensors2Data}
\alias{Sensors2Data-package}
\title{Derive data from measurements made with several ship-board sensors}
\description{
\pkg{Sensors2Data} generates metadata and derived data from
files generated by CTD, Marport Trawl Explorer, anr RBR Duet.
}
\details{
\emph{U.S. Geological Survey} (USGS) Computer Program \pkg{Sensors2Data}
  version 0.1.0.0000
Written by David M. Warner, USGS - Great Lakes Science Center
\href{http://www.glsc.usgs.gov/}{glsc.usgs.gov}, Ann Arbor, Michigan, USA.
Written in programming language R (R Core Team, 2019, www.R-project.org),
version 3.6.1 (2019-07-05).
Run on a PC with Intel(R) Core(TM) I7-76820HQ CPU, 2.90 GHz processor,
40.0 GB RAM, and Microsoft Windows 10 Professional operating system 2019.

Source code is available from David M. Warner on GitHub,
\href{https://github.com/dmwarn/Sensors2Data}{github.com/dmwarn/Sensors2Data},
\emph{dmwarner (at) usgs (dot) gov}.

\emph{Disclaimer:}  Although this program has been used by the USGS,
no warranty, expressed or implied, is made by the USGS or the United States
Government as to the accuracy and functioning of the program and related
program material nor shall the fact of distribution constitute any such
warranty, and no responsibility is assumed by the USGS in connection
therewith.
}
