% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeOnBottomTemp.R
\name{TimeOnBottom}
\alias{TimeOnBottom}
\title{Estimate time on bottom for bottom trawl}
\usage{
TimeOnBottom()
}
\arguments{
\item{fdep}{Numeric scale the user enters after being prompted which is equal to the desired
fishing depth in meters.}
}
\value{
An interactive graph from which the user can obtain time on bottom.
}
\description{
Estimate time on bottom for bottom trawl from RBR or other depth data
}
\details{
Create an interactive graph of depth data recorded by an instrument
 deployed on the headrope of a trawl. The user can draw a box around a subset
 of data and obtain the mean depth, start time, end time, and elapsed time of that
 subset.
}
